requirements.txt
natsort==8.4.0
numba==0.59.1
numpy==2.1.3
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
openpyxl==3.1.2
Pillow==9.2.0
Pillow==11.0.0
prefetch_generator==1.0.3
scipy==1.14.1
skimage==0.0
torch==2.0.1
torchsummary==1.5.1
torchvision==0.13.1
tqdm==4.65.0
